+++ exec.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"reflect"
-	"strings"
 
 	"github.com/lib/pq"
 	"golang.org/x/xerrors"
@@ -65,95 +64,6 @@ func setPrimaryKey(rv reflect.Value, id int64) {
 		err := fmt.Errorf("Unknown type to set primary key: %s", rv.Type())
 		panic(err)
 	}
-}
-
-// InsertBulk takes a table name and a slice of struct and inserts
-// the record in the DB with one Exec.
-// The given data needs to be:
-//
-// *[]*strcut
-// *[]struct
-// []*struct
-// []struct
-//
-// sqlpro will executes one INSERT statement per call.
-func (db *DB) InsertBulk(table string, data interface{}) error {
-	var (
-		rv         reflect.Value
-		structMode bool
-		err        error
-	)
-
-	rv, structMode, err = checkData(data)
-	if err != nil {
-		return err
-	}
-
-	if structMode {
-		return fmt.Errorf("InsertBulk: Need Slice to insert bulk.")
-	}
-
-	key_map := make(map[string]*fieldInfo, 0)
-	rows := make([]map[string]interface{}, 0)
-
-	if rv.Len() == 0 {
-		return nil
-	}
-
-	for i := 0; i < rv.Len(); i++ {
-		row := reflect.Indirect(rv.Index(i)).Interface()
-
-		values, structInfo, err := db.valuesFromStruct(row)
-
-		if err != nil {
-			return xerrors.Errorf("sqlpro.InsertBulk error: %w", err)
-		}
-
-		rows = append(rows, values)
-		for key := range values {
-			key_map[key] = structInfo[key]
-		}
-	}
-
-	insert := strings.Builder{} // make([]string, 0)
-	keys := make([]string, 0, len(key_map))
-
-	insert.WriteString("INSERT INTO ")
-	insert.WriteString(db.Esc(table))
-	insert.WriteString(" (")
-
-	idx := 0
-	for key := range key_map {
-		if idx > 0 {
-			insert.WriteRune(',')
-		}
-		insert.WriteString(db.Esc(key))
-		keys = append(keys, key)
-		idx++
-	}
-
-	insert.WriteString(") VALUES ")
-
-	for idx, row := range rows {
-		if idx > 0 {
-			insert.WriteRune(',')
-		}
-		insert.WriteRune('(')
-		for idx2, key := range keys {
-			if idx2 > 0 {
-				insert.WriteRune(',')
-			}
-			insert.WriteString(db.EscValueForInsert(row[key], key_map[key]))
-		}
-		insert.WriteRune(')')
-	}
-
-	_, err = db.DB.Exec(insert.String())
-	if err != nil {
-		return sqlError(err, insert.String(), []interface{}{})
-	}
-
-	return nil
 }
 
 func (db *DB) InsertBulkCopyIn(table string, data interface{}) error {
